Revert "Fix float conversion"
This reverts commit 9cf0c0e09a38e27309c35a59946135f7da59a81e.

+++ model/smart-impl/src/main/java/com/evolveum/midpoint/smart/impl/scoring/MappingsQualityAssessor.java
@@ -39,10 +39,10 @@ public class MappingsQualityAssessor {
         }
 
         if (totalShadowValues == 0) {
-            return (float) -1.0;
+            return -1.0f;
         }
         float quality = (float) matchedShadowValues / totalShadowValues;
-        return (float) (Math.round(quality * 100.0) / 100.0);
+        return Math.round(quality * 100.0f) / 100.0f;
     }
 
     private String applyExpression(ExpressionType expressionType, String input) {
